report_build_app/requirements.txt
click==7.1.2
docxtpl==0.11.1
Flask==1.1.2
install==1.3.4
itsdangerous==1.1.0
Jinja2==2.11.2
lxml==4.6.1
MarkupSafe==1.1.1
numpy==1.19.3
pandas==1.1.4
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-docx==0.8.10
pytz==2020.1
six==1.15.0
Werkzeug==1.0.1
